Bug fix, reorg, and refactor
Moved components into separate directories; using css module for MessageListHeader styling; fixed bug with start/stop button
