+++ src/components/MessageListHeader/style.module.css
@@ -1,4 +1,4 @@
-.message-list-header button {
+.button {
   width: 100px;
   margin: 8px;
   padding: 12px 0;
